Make sure the item has a width
When included on `kwc-nav`, the caret was being misplaced because it would calculate its position before the slotted content being loaded. This caused the flex box container ignore the existence of some items and messing with the offsets. By ensuring it will have a width it will calculate the caret position correctly. This is a problem that happened only on Safari.